refactor(api): Extract common anonymous auth
This extracts some constants for use in configuring anonymous
authentication.
